+++ R/layers-convolutional.R
@@ -465,7 +465,7 @@ layer_conv_3d_transpose <- function(object, filters, kernel_size, strides = c(1,
 #' @family convolutional layers 
 #'   
 #' @export
-layer_separable_conv_2d <- function(object, filters, kernel_size, strides = c(1L, 1L), padding = "valid", data_format = NULL, 
+layer_separable_conv_2d <- function(object, filters, kernel_size, strides = c(1, 1), padding = "valid", data_format = NULL, 
                                     depth_multiplier = 1L, activation = NULL, use_bias = TRUE, 
                                     depthwise_initializer = "glorot_uniform", pointwise_initializer = "glorot_uniform", bias_initializer = "zeros", 
                                     depthwise_regularizer = NULL, pointwise_regularizer = NULL, bias_regularizer = NULL, activity_regularizer = NULL, 
@@ -501,6 +501,51 @@ layer_separable_conv_2d <- function(object, filters, kernel_size, strides = c(1L
     weights = weights
   ))
   
+}
+
+#' Depthwise separable 2D convolution.
+#'
+#' Depthwise Separable convolutions consists in performing just the first step
+#' in a depthwise spatial convolution (which acts on each input channel
+#' separately). The `depth_multiplier` argument controls how many output
+#' channels are generated per input channel in the depthwise step.
+#' 
+#' @inheritParams layer_separable_conv_2d
+#' 
+#' @family convolutional layers
+#'
+#' @export
+layer_depthwise_conv_2d <- function(object, kernel_size, strides = c(1, 1), padding = "valid", depth_multiplier = 1,
+                                    data_format = NULL, activation = NULL, use_bias = TRUE,
+                                    depthwise_initializer = "glorot_uniform", bias_initializer = "zeros",
+                                    depthwise_regularizer = NULL, bias_regularizer = NULL, activity_regularizer = NULL,
+                                    depthwise_constraint = NULL, bias_constraint = NULL, input_shape = NULL,
+                                    batch_input_shape = NULL, batch_size = NULL, dtype = NULL, 
+                                    name = NULL, trainable = NULL, weights = NULL) {
+  
+  create_layer(keras$layers$DepthwiseConv2D, object, list(
+    kernel_size = as_integer_tuple(kernel_size),
+    strides = as_integer_tuple(strides),
+    padding = padding,
+    depth_multiplier = as.integer(depth_multiplier),
+    data_format = data_format,
+    activation = activation,
+    use_bias = use_bias,
+    depthwise_initializer = depthwise_initializer,
+    bias_initializer = bias_initializer,
+    depthwise_regularizer = depthwise_regularizer,
+    bias_regularizer = bias_regularizer,
+    activity_regularizer = activity_regularizer,
+    depthwise_constraint = depthwise_constraint,
+    bias_constraint = bias_constraint,
+    input_shape = normalize_shape(input_shape),
+    batch_input_shape = normalize_shape(batch_input_shape),
+    batch_size = as_nullable_integer(batch_size),
+    dtype = dtype,
+    name = name,
+    trainable = trainable,
+    weights = weights
+  ))
 }